attached_assets/requirements.txt
pymeeus==0.5.12
lunardate==0.2.0
pandas>=2.0.0
python-dateutil>=2.8.2
pytz>=2023.3
